+++ ChangeLog.md
@@ -7,6 +7,7 @@ ChangeLog
 ### Changed
 
 * Library: Circular buffer: Use size_type instead of size_t.
+* Library: Circular buffer: Add requirements to template parameters by using concepts.
 
 [0.0.0.2026021901] - 2026-02-19
 -------------------------------