+++ ChangeLog.md
@@ -9,6 +9,7 @@ ChangeLog
 * Library: Circular buffer: Use size_type instead of size_t.
 * Library: Circular buffer: Add requirements to template parameters by using concepts.
 * Library: Circular buffer: Remove template parameter `IndexT'.
+* Library: Circular buffer: Specify ordering requirements for atomic variables.
 
 [0.0.0.2026021901] - 2026-02-19
 -------------------------------