+++ ChangeLog.md
@@ -7,6 +7,7 @@ ChangeLog
 ### Changed
 
 * Library: Event loop toolbox: Remove `on_destroy_event_loop'.
+* Library: Event loop toolbox: Add data type `event_entry'.
 
 [0.0.0.2026022501] - 2026-02-25
 -------------------------------