+++ ChangeLog.md
@@ -1,6 +1,13 @@
 ChangeLog
 =========
 
+[0.0.0.2026092201] - 2026-09-22
+-------------------------------
+
+### Changed
+
+* Library: Mailbox: Make it possible to handle a stop_token object.
+
 [0.0.0.2026032201] - 2026-03-22
 -------------------------------